+++ banto/__init__.py
@@ -30,4 +30,4 @@ __all__ = [
     "KeyNotFoundError",
     "ProfileManager",
 ]
-__version__ = "4.2.0"
+__version__ = "5.0.0"